+++ tests/unit/lib/testutils.c
@@ -24,3 +24,17 @@ char os_secure_memcmp(const void *src1, const void *src2, size_t length)
     }
     return xoracc;
 }
+
+uint32_t cx_crc32_hw(const uint8_t *data, size_t len) {
+    uint32_t crc = ~0;
+    const uint8_t *end = data + len;
+
+    while (data < end) {
+        crc ^= *data++;
+        for (uint8_t i = 0; i < 8; i++) {
+            uint32_t mask = ~((crc & 1) - 1);
+            crc = (crc >> 1) ^ (0xEDB88320 & mask);
+        }
+    }
+    return ~crc;
+}